#6124: move code related to match check into utils package
